library/pyproject.toml
[tool.poetry]
name = "suiii"
version = "0.1.0"
description = ""
authors = ["AbdoGad <bedo2171311@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
djangorestframework = "^3.14.0"
markdown = "^3.4.1"
django-filter = "^22.1"
django = "^4.1.2"
django-extensions = "^3.2.1"
django-rest-knox = "^4.2.0"
knox = "^0.1.14"
jinja2 = "^3.1.2"
boto3 = "^1.26.3"
djangorestframework-simplejwt = "^5.2.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
